+++ Common/src/main/java/com/tristankechlo/whatdidijustkill/client/AbstractEntityToast.java
@@ -1,17 +1,18 @@
 package com.tristankechlo.whatdidijustkill.client;
 
+import com.mojang.blaze3d.systems.RenderSystem;
+import com.mojang.blaze3d.vertex.PoseStack;
 import com.tristankechlo.whatdidijustkill.config.types.ToastTheme;
-import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.GuiComponent;
 import net.minecraft.client.gui.components.toasts.Toast;
 import net.minecraft.client.gui.components.toasts.ToastComponent;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 
 public abstract class AbstractEntityToast implements Toast {
 
     private final Component firstLine; // not null
     private final Component secondLine; // might be null
-    protected ResourceLocation backgroundTexture = ToastTheme.ADVANCEMENT.getBackgroundTexture();
+    protected int backgroundTextureOffsetY = ToastTheme.ADVANCEMENT.getOffsetY();
     protected int displayTime = 2000;
     protected boolean textShadow = true;
 
@@ -21,23 +22,30 @@ public abstract class AbstractEntityToast implements Toast {
     }
 
     @Override
-    public Visibility render(GuiGraphics graphics, ToastComponent parent, long displayTime) {
+    public Visibility render(PoseStack poseStack, ToastComponent parent, long displayTime) {
         if (!ToastHandler.toastsEnabled) {
             return Visibility.HIDE;
         }
 
         // render background texture
-        graphics.blitSprite(this.backgroundTexture, 0, 0, this.width(), this.height());
+        RenderSystem.setShaderTexture(0, TEXTURE);
+        GuiComponent.blit(poseStack, 0, 0, 0, this.backgroundTextureOffsetY, this.width(), this.height());
 
         // draw entity texture
-        this.renderEntityImage(graphics);
+        this.renderEntityImage(poseStack);
 
         // draw text
         if (this.secondLine != null) {
-            graphics.drawString(parent.getMinecraft().font, this.secondLine, 30, 17, 16777215, this.textShadow);
+            if (this.textShadow) {
+                parent.getMinecraft().font.drawShadow(poseStack, this.secondLine, 30, 17, 16777215);
+            }
+            parent.getMinecraft().font.draw(poseStack, this.secondLine, 30, 17, 16777215);
         }
         int y = this.secondLine == null ? 12 : 7;
-        graphics.drawString(parent.getMinecraft().font, this.firstLine, 30, y, 16777215, this.textShadow);
+        if (this.textShadow) {
+            parent.getMinecraft().font.drawShadow(poseStack, this.firstLine, 30, y, 16777215);
+        }
+        parent.getMinecraft().font.draw(poseStack, this.firstLine, 30, y, 16777215);
 
         // remove toast when time is over
         return (double) displayTime >= this.displayTime * parent.getNotificationDisplayTimeMultiplier()
@@ -45,6 +53,6 @@ public abstract class AbstractEntityToast implements Toast {
                 : Visibility.SHOW;
     }
 
-    protected abstract void renderEntityImage(GuiGraphics graphics);
+    protected abstract void renderEntityImage(PoseStack graphics);
 
 }